ReservationService improved

+++ src/main/java/com/spring/booking/controller/ReservationController.java
@@ -33,7 +33,7 @@ public class ReservationController {
         return reservationService.getAvailableRooms(getAvailabilityDTO.getStartDate(), getAvailabilityDTO.getEndDate(), getAvailabilityDTO.getNumberOfPersons());
     }
     @GetMapping("/numberOfAvailableRooms")
-    public Integer getNumberOfAvailableRooms(@RequestBody AvailabilityOfHotelRoomsDTO availabilityOfHotelRoomsDTO){
+    public Long getNumberOfAvailableRooms(@RequestBody AvailabilityOfHotelRoomsDTO availabilityOfHotelRoomsDTO){
         return reservationService.getNumberOfAvailableRooms(availabilityOfHotelRoomsDTO.getStartDate(), availabilityOfHotelRoomsDTO.getEndDate(), availabilityOfHotelRoomsDTO.getIdHotel());
     }
     @GetMapping("/priceForAllReservations")
